requirements.txt
django==4.0.4
python-magic==0.4.25

